requirements.txt
Django==1.10.8
djangorestframework==3.7.7
psycopg2==2.6.2
